ci-vuln: renamed platform module to oqplatform to avoid conflicts with system platform package
