+++ openquakeplatform/openquakeplatform/test/__init__.py
@@ -2,7 +2,7 @@ import time
 import sys
 import nose
 
-from openquakeplatform.test.platform import Platform
+from openquakeplatform.test.oqplatform import Platform
 
 # uncomment to run pdb when untrapped exception is raised
 #
